Move EnvironmentSettingNames to infrastructure
