+++ source/TimeSeries/Infrastructure/EnvironmentSettingNames.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.TimeSeries.TimeSeriesBundleIngestor
+namespace Energinet.DataHub.TimeSeries.Infrastructure
 {
     public static class EnvironmentSettingNames
     {
